pyproject.toml
[project]
name = "ts_streaming"
version = "0.1.0"
description = "Tradesignals data processing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "quixstreams>=3.1.0",
    "websockets>=13.1",
]

[tool.uv.sources]

[tool.uv.workspace]
members = [
    "option_trades",
    "option_trades_sink",
    "option_trade_aggs",
]
